backend/pyproject.toml
[project]
name = "finance-planner"
version = "0.1.0"
description = "Personal financial planning tool with Monte Carlo simulation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
